Rename llama_completion to llama_completer.py, LlamaCompleter to LlamaGRPOCompleter
